pyproject.toml
[project]
name = "ml"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-genai>=1.19.0",
    "google-generativeai>=0.8.5",
    "opencv-python>=4.11.0.86",
    "pandas>=2.3.0",
    "praw>=7.8.1",
    "python-dotenv>=1.1.0",
    "tqdm>=4.67.1",
]
